minor tweaks
Made a modification to the profile panel to make fighter abilities easier to see.
Added post process to the lobbies.
Added post process to the fields.
Final adjustments to the profile screen on mobile.
Created a script to load colourgrading across all scenes.
Modified a script to change instructions depending on platform.